backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tradebrain-backend"
version = "0.1.0"
description = "TradeBrain backend MVP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "ib_async>=2.1,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

